refactor(format): move summary JSON writers
